### 2018-01-03 #179 @GuangchuangYu
treeio for parsing and exporting phylogenetic trees
Parsing and exporting phylogenetic trees with associated data.

```
Package: treeio
Title: Base Classes and Functions for Phylogenetic Tree Input and Output
Version: 1.3.10
    person("Guangchuang", "Yu",     email = "guangchuangyu@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-6485-8781")),
    person("Tommy Tsan-Yuk", "Lam", email = "tylam.tommy@gmail.com",   role = c("aut", "ths")),
    person("Casey", "Dunn",         email = "casey_dunn@brown.edu",    role = "ctb"),
    person("Bradley", "Jones",      email = "brj1@sfu.ca",             role = "ctb")
	  )
Description: Base classes and functions for parsing and exporting phylogenetic trees. 'treeio' supports parsing analysis findings from commonly used software packages, allows linking exteranl data to phylogeny and merging tree data obtained from different sources. It also supports exporting phylogenetic tree with heterogeneous associated data to a single tree file.
    ape,
    rvcheck,
    tidytree (>= 0.1.6)
    ggtree,
    prettydoc,
    tidyr
License: Artistic-2.0
biocViews: Software, Annotation, Clustering, DataImport, DataRepresentation,
    Alignment, MultipleSequenceAlignment
```

<https://github.com/GuangchuangYu/treeio>

data extraction

[e.g., "data extraction, because the package parses a scientific data file format"]

Researchers doing ecology and evolution.

No.

No.

	- [x] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

No.
