+++ README.md
@@ -4,13 +4,40 @@ treeio: Base classes and functions for phylogenetic tree input and output
 
 <img src="https://raw.githubusercontent.com/Bioconductor/BiocStickers/master/treeio/treeio.png" height="200" align="right" />
 
-[![releaseVersion](https://img.shields.io/badge/release%20version-1.2.1-green.svg?style=flat)](https://bioconductor.org/packages/treeio) [![develVersion](https://img.shields.io/badge/devel%20version-1.3.10-green.svg?style=flat)](https://github.com/guangchuangyu/treeio) [![Bioc](http://www.bioconductor.org/shields/years-in-bioc/treeio.svg)](https://www.bioconductor.org/packages/devel/bioc/html/treeio.html#since) [![total](https://img.shields.io/badge/downloads-4958/total-blue.svg?style=flat)](https://bioconductor.org/packages/stats/bioc/treeio) [![month](https://img.shields.io/badge/downloads-903/month-blue.svg?style=flat)](https://bioconductor.org/packages/stats/bioc/treeio)
+[![releaseVersion](https://img.shields.io/badge/release%20version-1.2.1-green.svg?style=flat)](https://bioconductor.org/packages/treeio) [![develVersion](https://img.shields.io/badge/devel%20version-1.3.10-green.svg?style=flat)](https://github.com/guangchuangyu/treeio) [![Bioc](http://www.bioconductor.org/shields/years-in-bioc/treeio.svg)](https://www.bioconductor.org/packages/devel/bioc/html/treeio.html#since) [![total](https://img.shields.io/badge/downloads-5146/total-blue.svg?style=flat)](https://bioconductor.org/packages/stats/bioc/treeio) [![month](https://img.shields.io/badge/downloads-691/month-blue.svg?style=flat)](https://bioconductor.org/packages/stats/bioc/treeio)
 
-[![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active) [![codecov](https://codecov.io/gh/GuangchuangYu/treeio/branch/master/graph/badge.svg)](https://codecov.io/gh/GuangchuangYu/treeio) [![Last-changedate](https://img.shields.io/badge/last%20change-2017--12--20-green.svg)](https://github.com/GuangchuangYu/treeio/commits/master) [![GitHub forks](https://img.shields.io/github/forks/GuangchuangYu/treeio.svg)](https://github.com/GuangchuangYu/treeio/network) [![GitHub stars](https://img.shields.io/github/stars/GuangchuangYu/treeio.svg)](https://github.com/GuangchuangYu/treeio/stargazers)
+[![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active) [![codecov](https://codecov.io/gh/GuangchuangYu/treeio/branch/master/graph/badge.svg)](https://codecov.io/gh/GuangchuangYu/treeio) [![Last-changedate](https://img.shields.io/badge/last%20change-2018--01--03-green.svg)](https://github.com/GuangchuangYu/treeio/commits/master) [![GitHub forks](https://img.shields.io/github/forks/GuangchuangYu/treeio.svg)](https://github.com/GuangchuangYu/treeio/network) [![GitHub stars](https://img.shields.io/github/stars/GuangchuangYu/treeio.svg)](https://github.com/GuangchuangYu/treeio/stargazers)
 
[…]
