### 2018-01-09 #179 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @GuangchuangYu  !!

- I do believe Artistic is a CRAN accepted license and a OSI accepted license, can you check that box?
- It looks like you have continuous integration with Bioconductor and Appveyor, so can check that box as well.
- I see you checked the submit to MEE box. This is our first one!
- You can put a ropensci review badge in your README

- Goodpractice output for you and reviewers to consider:

```r
It is good practice to

    R/AllGenerics.R:12:NA
    R/AllGenerics.R:26:NA
    R/ancestor.R:5:NA
    R/ancestor.R:6:NA
    R/ancestor.R:7:NA
    ... and 1326 more lines

  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/hyphy.R:205:18
    R/hyphy.R:207:18
    R/jtree.R:66:20
    R/jtree.R:68:21
    R/jtree.R:69:22
    ... and 50 more lines

    R/AllGenerics.R:26:1
    R/beast.R:136:1
    R/beast.R:199:1
    R/beast.R:206:1
    R/beast.R:239:1
    ... and 45 more lines

    vapply() instead.

    R/beast.R:71:14
    R/beast.R:97:16
    R/beast.R:233:21
    R/beast.R:239:19
    R/beast.R:279:11
    ... and 16 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/method-groupOTU.R:61:19
    R/write-beast.R:88:25
    R/write-beast.R:160:15
    tests/testthat/test-nhx.R:75:25
```

During check:

```r
* checking top-level files ... NOTE
Non-standard file/directory found at top level:
  ‘CONTRIBUTORS.md’
```

Seeking reviewers now 🕐

---

Reviewers:  xx xx
Due date: xxxx
