### 2018-01-10 #179 @GuangchuangYu
thanks @sckott !

I have checked all the boxes you mentioned and add ropensci badge on README.md.
