### 2018-01-10 3f872c6
clean up code according to goodpractice
