+++ R/reexport.R
@@ -10,6 +10,25 @@ tidytree::get.fields
 ##' @export
 tidytree::as.treedata
 
+##' @importFrom tidytree ancestor
+##' @export
+tidytree::ancestor
+
+##' @importFrom tidytree parent
+##' @export
+tidytree::parent
+
+##' @importFrom tidytree child
+##' @export
+tidytree::child
+
+##' @importFrom tidytree offspring
+##' @export
+tidytree::offspring
+
+##' @importFrom tidytree rootnode
+##' @export
+tidytree::rootnode
 
 ##' @importFrom dplyr full_join
 ##' @export
