### 2018-01-10 #179 @GuangchuangYu
## current goodpractice output

```
It is good practice to

    R/beast.R:23:NA
    R/beast.R:24:NA
    R/beast.R:25:NA
    R/beast.R:26:NA
    R/beast.R:81:NA
    ... and 911 more lines

    tests/testthat/test-r8s.R:4:1
```
