+++ tests/testthat/test-fulljoin.R
@@ -1,0 +1,12 @@
+context("fulljoin")
+
+
+rstfile <- system.file("extdata/PAML_Baseml", "rst", package="treeio")
+tree <- read.paml_rst(rstfile)
+x <- data_frame(label = as.phylo(tree)$tip.label, trait = rnorm(Ntip(tree)))
+y <- full_join(tree, x, by="label")
+
+test_that("linking external data to treedata", {
+    expect_true(is(y, "treedata"))
+    expect_true("trait" %in% get.fields(y))
+})
