+++ R/raxml2nwk.R
@@ -10,7 +10,7 @@
 ##' @author Guangchuang Yu
 raxml2nwk <- function(infile, outfile="raxml.tree") {
     raxml <- read.raxml(infile)
-    nlabel <- raxml@data[,2]
+    nlabel <- raxml@data[[2]]
     nlabel[is.na(nlabel)] <- ""
     raxml@phylo$node.label <- nlabel
     write.tree(raxml@phylo, file=outfile)
