+++ man/write.jplace.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/jtree.R
+% Please edit documentation in R/defunct.R
 \name{write.jplace}
 \alias{write.jplace}
 \title{write.jplace}
