### 2018-01-12 bc87638
more dependencies for travis
