### 2018-01-19 efa7791
bug fixed of converting phylo4d
