### 2018-02-16 c7e8bc9
fixed a bug in as.phylo.phylo4
