### 2018-02-19 #179 @fmichonneau
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

* The README could be edited to be made easier to read for beginners. Someone
  new to R might find it difficult to understand what this package is about.

* The vignettes are great. It might be helpful though to better document what to
  do with the trees/objects created by the `treeio` functions. I think some of
  this information is included in the vignettes of the `ggtree` package, so
  referring to these vignettes here could be helpful.

* The function documentation are not very detailed. It is also not clear where
  to get information about the structure of the `treedata` objects returned by
  the `read.*` functions, available methods to extract information they contain,
  etc.

* Community guidelines: the repository doesn't have contribution guidelines (no
  section in README or a CONTRIBUTING file).

During installation I see:

```
** byte-compile and prepare package for lazy loading
Note: wrong number of arguments to '>'
```

`install_github` also returns "Warning message: this repository contains submodules, may not function as expected." A note in the README about whether this warning is safe to ignore would be useful.

It also seems that the order in which `treeio` and `ggtree` is installed matters. I haven't tested this much, but if it's indeed the case, some indication in which order to install these packages, to avoid issues down the road would be useful.

In the submission, the author indicates that no other packages have similar
functionalities, but there are actually many packages that provide ways to
import phylogenetic data in R. `treeio` attempts to be comprehensive by pulling functionalities from other packages, and re-using code, but
many of the packages listed under the section "Getting trees into R" from the
"Phylogenetics" CRAN task view overlap with this package
https://cran.r-project.org/view=Phylogenetics

In general, it seems that an overview of where `treeio` fits in the phylogenetics landscape in R would be welcome.

It looks like there are extensive tests for the code, but the codecov badge is currently broken. A quick look at the log files on Travis reveal that it's because the patch version for the minimum R version is listed in the README (it should be `R (>= 3.4.0)` instead of `R (>= 3.4.2)`), this is a new requirement in R-devel.

* typo in Description and README: "exteranl" -> "external"
* the link to "Base Classes and Functions for Phylogenetic Tree Input and
  Output" in https://guangchuangyu.github.io/treeio/documentation/ is broken.
