### 2018-02-19 #179 @sckott
thanks for your review @fmichonneau !  can you give an estimate of how many hrs spent on the review?
