### 2018-02-23 80d5c6f
update generic function definition
