+++ DESCRIPTION
@@ -7,7 +7,7 @@ Authors@R: c(
     person("Casey", "Dunn",         email = "casey_dunn@brown.edu",    role = "ctb"),
     person("Bradley", "Jones",      email = "brj1@sfu.ca",             role = "ctb")
 	  )
-Description: Base classes and functions for parsing and exporting phylogenetic trees. 'treeio' supports parsing analysis findings from commonly used software packages, allows linking exteranl data to phylogeny and merging tree data obtained from different sources. It also supports exporting phylogenetic tree with heterogeneous associated data to a single tree file.
+Description: Base classes and functions for parsing and exporting phylogenetic trees. 'treeio' supports parsing analysis findings from commonly used software packages, allows linking external data to phylogeny and merging tree data obtained from different sources. It also supports exporting phylogenetic tree with heterogeneous associated data to a single tree file.
 Depends:
     R (>= 3.4.2)
 Imports:
