+++ DESCRIPTION
@@ -9,7 +9,7 @@ Authors@R: c(
 	  )
 Description: Base classes and functions for parsing and exporting phylogenetic trees. 'treeio' supports parsing analysis findings from commonly used software packages, allows linking external data to phylogeny and merging tree data obtained from different sources. It also supports exporting phylogenetic tree with heterogeneous associated data to a single tree file.
 Depends:
-    R (>= 3.4.2)
+    R (>= 3.4.0)
 Imports:
     ape,
     dplyr,
