### 2018-02-26 #179 @GuangchuangYu
thanks for the comments. I will address these issues asap.
