### 2018-02-26 739d8b5
address ropensci comments
