+++ README.Rmd
@@ -36,14 +36,28 @@ library("yyplot")
 
 [![platform](http://www.bioconductor.org/shields/availability/devel/treeio.svg)](https://www.bioconductor.org/packages/devel/bioc/html/treeio.html#archives)
 [![Build Status](http://www.bioconductor.org/shields/build/devel/bioc/treeio.svg)](https://bioconductor.org/checkResults/devel/bioc-LATEST/treeio/)
-[![Linux/Mac Travis Build Status](https://img.shields.io/travis/GuangchuangYu/treeio/master.svg?label=Mac%20OSX%20%26%20Linux)](https://travis-ci.org/GuangchuangYu/treeio)
+[![Linux Travis Build Status](https://img.shields.io/travis/GuangchuangYu/treeio/master.svg?label=Linux)](https://travis-ci.org/GuangchuangYu/treeio)
 [![AppVeyor Build Status](https://img.shields.io/appveyor/ci/Guangchuangyu/treeio/master.svg?label=Windows)](https://ci.appveyor.com/project/GuangchuangYu/treeio)
-[![Twitter](https://img.shields.io/twitter/url/https/github.com/GuangchuangYu/treeio.svg?style=social)](https://twitter.com/intent/tweet?hashtags=treeio&url=http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12628/abstract&screen_name=guangchuangyu)
+[![Twitter](https://img.shields.io/twitter/url/http/shields.io.svg?style=social&logo=twitter)](https://twitter.com/intent/tweet?hashtags=treeio&url=http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12628/abstract&screen_name=guangchuangyu)
 
 
-```{r comment="", echo=FALSE, results='asis'}
+```{r comment="", echo=FALSE, results='asis', eval=FALSE}
 cat(packageDescription('treeio')$Description)
 ```
+
+`treeio` is an R package to make it easier to import and store phylogenetic tree
+with associated data; and to link external data from different sources to
+phylogeny. It also supports exporting phylogenetic tree with heterogeneous
+associated data to a single tree file and can be served as a platform for
+merging tree with associated data and converting file formats.
+
+See the `treeio` project website, <https://guangchuangyu.github.io/treeio>, and
+package vignette for more details.
+
+### Vignettes
+
++ [Importing trees with data](http://bioconductor.org/packages/devel/bioc/vignettes/treeio/inst/doc/Importer.html)
++ [Exporting trees with data](http://bioconductor.org/packages/devel/bioc/vignettes/treeio/inst/doc/Exporter.html)
 
 
 ## Authors
@@ -70,16 +84,6 @@ Or the development version from github:
 devtools::install_github("GuangchuangYu/treeio")
 ```
 
-
-## Vignettes
-
-+ [Importing trees with data](http://bioconductor.org/packages/devel/bioc/vignettes/treeio/inst/doc/Importer.html)
-+ [Exporting trees with data](http://bioconductor.org/packages/devel/bioc/vignettes/treeio/inst/doc/Exporter.html)
-
-
-For details, please visit our project website, <https://guangchuangyu.github.io/treeio>.
-
-
 ## Download stats
 
 `r badge_download_bioc("treeio")`
@@ -99,4 +103,8 @@ ggsave(p, file="docs/images/dlstats.png", width=8, height=4)
 <img src="docs/images/dlstats.png" width="890"/>
 
 
+## Contributing
+
+We welcome any contributions! By participating in this project you agree to
+abide by the terms outlined in the [Contributor Code of Conduct](CONDUCT.md).
 
