+++ site_src/content/documentation/index.rmd
@@ -9,7 +9,9 @@ weight: 10
 
 ## <i class="fa fa-book"></i> Vignettes
 
-+ [Base Classes and Functions for Phylogenetic Tree Input and Output](https://bioconductor.org/packages/devel/bioc/vignettes/treeio/inst/doc/treeio.html)
+
++ [Importing trees with data](http://bioconductor.org/packages/devel/bioc/vignettes/treeio/inst/doc/Importer.html)
++ [Exporting trees with data](http://bioconductor.org/packages/devel/bioc/vignettes/treeio/inst/doc/Exporter.html)
 + [ggtree for tree visualization and annotation](https://bioconductor.org/packages/devel/bioc/vignettes/ggtree/inst/doc/ggtree.html)
 
 
