### 2018-02-28 496fa2b
read.beast now compatible with taxa label contains ', " and space
