### 2018-02-28 #179 @GuangchuangYu
@uyedaj Thanks for your comments and here is my response.

>This package covers an important and useful niche in the R phylogenetics ecosystem that greatly improves parsing tree formats from different phylogenetics software packages. It will no doubt be useful for researchers, and I will be happy to use it in my own work in the future.

>The package has very clear and useful vignettes using provided datasets that make it clear how the package is used. This is done at the expense of examples in the .Rd files, which are largely lacking. The package is seamlessly integrated into the packages tidytree and ggtree. Most of the core functions exported in the package are directly related to reading in different types of source data. Other functions are exported from other packages. The package appears to pass all checks and tests included.

>When learning how to use the package, I immediately typed in the name of the object and noticed the
..@tree when the object is printed to the screen. However, this is not the name of the slot in the
S4 treedata object, which is instead ..@phylo; which seems unnecessarily confusing.

Had been changed to `..@phylo` in `tidytree >=0.1.7`.

>The provided datasets are a bit hefty for an R package at 3.2MB. Perhaps understandably however, but
the authors could see what they could do to try to get this down to a lower value (usually CRAN prefers
extdata to be than 1 MB.)

Since `treeio` supports several formats and I would like to present example for each one of them, there are several software outputs packed with this package.

One solution I can think of is to reduce the size by `gzip` the files at the expense of they can't be opened with text editor directly. What do you think?

>There is a typo in Exporter.Rmd
Exporting the tree structure is easy, user can use `as.phyo` method

corrected.

>What kind of `multiphylo` support is there in this package? I see that you can import a whole set of trees
using either `read.mrbayes` and `read.beast` but it tended to choke on average sized tree files when I
tested it. Perhaps the vignettes could show how the `beastList` class works and what it might be used for? I'm assuming that there is no way to associate node/tip data to a `beastList` object? If the package claims to read Beast and MrBayes formatted tree output, it seems like it should be prepared to read in posteriors of trees, or at least, specify clearly that these are not the primary functionality of the package.

`multiPhylo` can be used to stored tree structures of MCMC trees via `ape::read.nexus("mcmc_tree_file")`. However it is not designed for storing tree associated data (e.g. posterior values). The `beastList` object is just a class attribute (S3) without formal class definition. It is actually a list of the `treedata` object, which of course can read posteriors and other attributes of trees. I didn't demonstrate such functionality since it is currently not mature, and still have room to improve, especially performance.

>Overall, I find little to criticize in this excellent package!
