### 2018-02-28 #179 @GuangchuangYu
Thanks for @fmichonneau 's comments and here is my response.

### Documentation

>+ The README could be edited to be made easier to read for beginners. Someone
    new to R might find it difficult to understand what this package is about.

have updated and should be more easy to understand.

>+ The vignettes are great. It might be helpful though to better document what to
    do with the trees/objects created by the treeio functions. I think some of
    this information is included in the vignettes of the ggtree package, so
    referring to these vignettes here could be helpful.

`Visualizing tree data with ggtree` session was added in the `Importer` vignette.

>+ The function documentation are not very detailed. It is also not clear where
    to get information about the structure of the treedata objects returned by
    the read.* functions, available methods to extract information they contain,
    etc.

A session of `Getting information from treedata object` in the `Importer` vignette was added.

>+ Community guidelines: the repository doesn't have contribution guidelines (no
    section in README or a CONTRIBUTING file).

`Contributing` session was added in `README`.

### Installation

>During installation I see:
>
```
** byte-compile and prepare package for lazy loading
Note: wrong number of arguments to '>'
```

>`install_github` also returns "Warning message: this repository contains submodules, may not function as expected." A note in the README about whether this warning is safe to ignore would be useful.

All these warning message had been removed.

>It also seems that the order in which `treeio` and `ggtree` is installed matters. I haven't tested this much, but if it's indeed the case, some indication in which order to install these packages, to avoid issues down the road would be useful.

`treeio` suggest `ggtree` as it can enhance `treeio` by utilizing data imported by `treeio` to be visualized on a tree, while `ggtree` imports `treeio`'s tree data parsing functions.

`treeio` can work without `ggtree` while `ggtree` needs `treeio` to install.

When installing `ggtree`, the `treeio` package will be installed automatically. There should be no confusing here. Please correct me if I was wrong.

### Functionality

>In the submission, the author indicates that no other packages have similar
functionalities, but there are actually many packages that provide ways to
import phylogenetic data in R. treeio attempts to be comprehensive by pulling functionalities from other packages, and re-using code, but
many of the packages listed under the section "Getting trees into R" from the
"Phylogenetics" CRAN task view overlap with this package
https://cran.r-project.org/view=Phylogenetics

>In general, it seems that an overview of where `treeio` fits in the phylogenetics landscape in R would be welcome.

Several packages listed in `Getting trees into R` session from <https://cran.r-project.org/view=Phylogenetics> are for getting trees from online repositories.

`ape` and `phylobase` can only import Newick and Nexus. `RNeXML` is designed for `NeXML`. For software output, `simmap` is supported by `phyext2` and `phytools`. BEAST and MrBayes outputs are not full supported by `PHYLOCH` as you can refer to below: (only attributes of internal nodes can be parsed)

```
> file <- system.file("extdata/BEAST", "beast_mcc.tree", package="treeio")
> pb <- phyloch::read.beast(file)
> tb <- treeio::read.beast(file)
> attributes(pb)
$names
 [1] "edge"               "edge.length"        "Nnode"
 [4] "tip.label"          "rate_range_MIN"     "rate_range_MAX"
 [7] "length_range_MIN"   "length_range_MAX"   "length_95%_HPD_MIN"
[10] "length_95%_HPD_MAX" "height"             "rate"
[13] "rate_median"        "length"             "length_median"
[16] "rate_95%_HPD_MIN"   "rate_95%_HPD_MAX"   "height_95%_HPD_MIN"
[19] "height_95%_HPD_MAX" "height_range_MIN"   "height_range_MAX"
[22] "height_median"      "posterior"

$class
[1] "phylo"

$origin
[1] "/Library/R/library/treeio/extdata/BEAST/beast_mcc.tree"

> pb$rate
 [1]          NA 0.002922564 0.002907438 0.002909881 0.002887185 0.002894850
 [7] 0.002898022 0.002889015 0.002893996 0.002894633 0.002909435 0.002955187
[13] 0.002888937 0.002897986
> tb[["rate"]]
 [1] 0.002920138 0.002904671 0.002907438 0.002924273 0.002922564 0.002903580
 [7] 0.002904168 0.002912649 0.002931375 0.002910248 0.002908179 0.002924622
[13] 0.002955187 0.002909435 0.002894633 0.002896674 0.002893996 0.002926319
[19] 0.002889015 0.002897274 0.002923289 0.002930741 0.002897986 0.002888937
[25] 0.002898022 0.002894850 0.002887185 0.002909881          NA
```

There are many other formats that can not be imported directly into R and we developed `treeio` to address this issue. I have added an overview in the `Importer` vignette.

### Tests

>It looks like there are extensive tests for the code, but the codecov badge is currently broken. A quick look at the log files on Travis reveal that it's because the patch version for the minimum R version is listed in the README (it should be `R (>= 3.4.0)` instead of `R (>= 3.4.2)`), this is a new requirement in R-devel.

fixed and the codecov badge is available.

### Minor comments

>+ typo in Description and README: "exteranl" -> "external"

Corrected.

>+ the link to "Base Classes and Functions for Phylogenetic Tree Input and
    Output" in <https://guangchuangyu.github.io/treeio/documentation/> is broken.

Corrected.
