### 2018-02-28 #179 @sckott
thanks for your responses to reviewers @GuangchuangYu

@uyedaj @fmichonneau let me know if  you are happy with the maintainers responses - or feel free to keep the conversation going if you have further thoughts
