### 2018-03-06 #179 @sckott
@GuangchuangYu both reviewers are okay with changes. I'll have a quick look myself now to see if there's any lingering issues.
