### 2018-03-06 #179 @sckott
With the bioc versions of `treeio` and `ggtree` everything works fine, but with dev version of `treeio`, everything installs, but some tests are failing I think because some methods have been removed in `treeio`

```r
test-conversion.R:4: error: (unknown)
object ‘groupClade’ is not exported by 'namespace:treeio'
1: ggtree::ggtree at /Users/sckott/github/ropensci/treeio/tests/testthat/test-conversion.R:4
2: getExportedValue(pkg, name)
3: asNamespace(ns)
4: getNamespace(ns)
5: tryCatch(loadNamespace(name), error = function(e) stop(e))
6: tryCatchList(expr, classes, parentenv, handlers)
7: tryCatchOne(expr, names, parentenv, handlers[[1L]])
8: value[[3L]](cond)

test-treedata-accessor.R:34: error: (unknown)
object ‘groupClade’ is not exported by 'namespace:treeio'
1: ggtree::ggtree at /Users/sckott/github/ropensci/treeio/tests/testthat/test-treedata-accessor.R:34
2: getExportedValue(pkg, name)
3: asNamespace(ns)
4: getNamespace(ns)
5: tryCatch(loadNamespace(name), error = function(e) stop(e))
6: tryCatchList(expr, classes, parentenv, handlers)
7: tryCatchOne(expr, names, parentenv, handlers[[1L]])
8: value[[3L]](cond)
```

<details> <summary><strong>Session Info</strong></summary>

```r
Session info ------------------------------------------------------------------
 setting  value
 version  R version 3.4.4 beta (2018-03-05 r74359)
 system   x86_64, darwin15.6.0
 ui       X11
 language (EN)
 collate  en_US.UTF-8
 tz       America/Los_Angeles
 date     2018-03-06

Packages ----------------------------------------------------------------------
 package    * version    date       source
 ape          5.0        2017-10-30 CRAN (R 3.4.2)
 assertthat   0.2.0      2017-04-11 CRAN (R 3.4.0)
 base       * 3.4.4      2018-03-06 local
 bindr        0.1        2016-11-13 CRAN (R 3.4.0)
 bindrcpp     0.2        2017-06-17 CRAN (R 3.4.0)
 commonmark   1.4        2017-09-01 cran (@1.4)
 compiler     3.4.4      2018-03-06 local
 datasets   * 3.4.4      2018-03-06 local
 devtools     1.13.5     2018-02-18 CRAN (R 3.4.3)
 digest       0.6.15     2018-01-28 CRAN (R 3.4.3)
 dplyr        0.7.4      2017-09-28 CRAN (R 3.4.2)
 glue         1.2.0      2017-10-29 CRAN (R 3.4.2)
 graphics   * 3.4.4      2018-03-06 local
 grDevices  * 3.4.4      2018-03-06 local
 grid         3.4.4      2018-03-06 local
 jsonlite     1.5        2017-06-01 CRAN (R 3.4.0)
 lattice      0.20-35    2017-03-25 CRAN (R 3.4.4)
 lazyeval     0.2.1      2017-10-29 cran (@0.2.1)
 magrittr     1.5        2014-11-22 CRAN (R 3.4.0)
 memoise      1.1.0      2018-02-27 Github (hadley/memoise@611cfad)
 methods    * 3.4.4      2018-03-06 local
 nlme         3.1-131.1  2018-02-16 CRAN (R 3.4.4)
 parallel     3.4.4      2018-03-06 local
 pillar       1.2.1      2018-02-27 CRAN (R 3.4.3)
 pkgconfig    2.0.1      2017-03-21 CRAN (R 3.4.0)
 R6           2.2.2      2017-06-17 CRAN (R 3.4.0)
 Rcpp         0.12.15    2018-01-20 CRAN (R 3.4.3)
 rlang        0.2.0.9000 2018-03-06 Github (tidyverse/rlang@9ea33dd)
 roxygen2     6.0.1      2017-02-06 CRAN (R 3.4.0)
 rstudioapi   0.7        2017-09-07 CRAN (R 3.4.1)
 rvcheck      0.0.9      2017-07-10 cran (@0.0.9)
 stats      * 3.4.4      2018-03-06 local
 stringi      1.1.6      2017-11-17 CRAN (R 3.4.2)
 stringr      1.3.0.9000 2018-03-06 Github (tidyverse/stringr@097c0c0)
 testthat   * 2.0.0      2017-12-13 CRAN (R 3.4.3)
 tibble       1.4.2      2018-01-22 CRAN (R 3.4.3)
 tidytree     0.1.7      2018-02-27 CRAN (R 3.4.3)
 tools        3.4.4      2018-03-06 local
 treeio     * 1.3.12     <NA>       Bioconductor
 utils      * 3.4.4      2018-03-06 local
 withr        2.1.1.9000 2018-03-05 Github (jimhester/withr@5d05571)
 xml2         1.2.0      2018-01-24 CRAN (R 3.4.3)
```
