### 2018-03-07 #179 @GuangchuangYu
@sckott The current bioc dev branch of both treeio and ggtree is identical to github version.

The reason for the issue is that the OSX tarball of ggtree in Bioc dev is not updated due to the suggested package `ggimage` that is [not installed on their build server](http://bioconductor.org/checkResults/devel/bioc-LATEST/ggtree/merida2-checksrc.html).

The `ggimage` is suggested as it enhances `ggtree` for [annotating tree with images files](https://guangchuangyu.github.io/software/ggtree/vignettes/ggtree-ggimage.html).

The issue you presented here is on the Bioconductor side as `ggimage` is indeed available via CRAN.
