### 2018-03-08 #179 @GuangchuangYu
```r
install.packages("https://bioconductor.org/packages/devel/bioc/src/contrib/treeio_1.3.12.tar.gz", repo=NULL)
install.packages("https://bioconductor.org/packages/devel/bioc/src/contrib/ggtree_1.11.6.tar.gz", repo=NULL)
```

After I installed the treeio and ggtree using the source tarball on Bioconductor, I clone the repo and run the `rcmdcheck`. Everything just goes fine.

```
$ git clone https://git.bioconductor.org/packages/treeio
$ git clone https://git.bioconductor.org/packages/ggtree
$ cd treeio
$ make
...
── 0 errors ✔ | 0 warnings ✔ | 0 notes ✔

$ cd ggtree
$ make
...
── 0 errors ✔ | 0 warnings ✔ | 1 note ✖
```

The error you presented here should be due to inconsistent versions (i.e. mix installation of release and devel branches), since there is a lot of change in the `treeio` dev branch and also some codes were move to the `tidytree` package.
