### 2018-03-11 #179 @sckott
Thanks, works well.

----

Approved!  Thanks again for your submission @GuangchuangYu

* At this point we usually have maintainers transfer their package  to `ropensci` - bit I'm pretty sure that's not possible with Bioc pkgs, correct? We've had one other Bioc pkg (`genbankr`) and we planned to simply set up a mirror in our ropensci org account, but haven't got around to it yet - but that's the plan anyway for all Bioc pkgs
* Make sure to
    * add bioc/rOpenSci footer to README, see eg https://github.com/gmbecker/genbankr#package-affiliations
* In your `.github` dir can you also add a CONTRIBUTING.md and a PR template, see https://github.com/ropensci/dotgithubfiles/ for examples
* We're starting to roll out software metadata files to all ropensci pkgs via the Codemeta initiative, see  https://github.com/ropensci/codemetar/#codemetar  for how to include it in your pkg, after installing the pkg - should be easy as running `codemetar::write_codemeta()` in the root of your pkg
* For the MEE paper, go ahead now and submit your manuscript to MEE if you haven't already. They tell me there's an option to say the software has been reviewed by ropensci in their submission template

---

* Are you interested in doing a blog post for our blog https://ropensci.org/blog/ ?  either a short-form intro to it (https://ropensci.org/technotes/) or long-form post with more narrative about its development (https://ropensci.org/blog/). If so, we'll have our community manager @stefaniebutland get in touch with you on that
