+++ NEWS.md
@@ -1,3 +1,10 @@
+# treeio 1.3.13
+
++ mv project website to <https://guangchuangyu.github.io/software/treeio>
++ update for rOpenSci acceptance
+  - <https://github.com/ropensci/onboarding/issues/179#issuecomment-372127781>
+
+
 # treeio 1.3.12
 
 + read.beast now compatible with taxa label contains ', " and space (2018-02-27,
