+++ .github/CONTRIBUTING.md
@@ -1,0 +1,30 @@
+# CONTRIBUTING #
+
+### Bugs?
+
+* Submit an issue on the [Issues page](https://github.com/GuangchuangYu/treeio/issues)
+
+### Issues and Pull Requests
+
+If you are considering a pull request, you may want to open an issue first to discuss with the maintainer(s).
+
+### Code contributions
+
+* Fork this repo to your GitHub account
+* Clone your version on your account down to your machine from your account, e.g,. `git clone https://github.com/<yourgithubusername>/{repo}.git`
+* Make sure to track progress upstream (i.e., on our version of `treeio` at `GuangchuangYu/treeio`) by doing `git remote add upstream https://github.com/GuangchuangYu/treeio.git`. Before making changes make sure to pull changes in from upstream by doing either `git fetch upstream` then merge later or `git pull upstream` to fetch and merge in one step
+* Make your changes (bonus points for making changes on a new feature branch - see <https://guides.github.com/introduction/flow/> for how to contribute by branching, making changes, then submitting a pull request)
+* Push up to your account
+* Submit a pull request to home base (likely master branch, but check to make sure) at `GuangchuangYu/treeio`
+
+### Discussion forum
+
+Check out our [discussion forum](https://discuss.ropensci.org) if you think your issue requires a longer form discussion.
+
+### Prefer to Email?
+
+Email the person listed as maintainer in the `DESCRIPTION` file of this repo.
+
+Though note that private discussions over email don't help others - of course email is totally warranted if it's a sensitive problem of any kind.
+
+### Thanks for contributing!
