+++ site_src/content/documentation/index.rmd
@@ -1,24 +1,0 @@
----
-date: 2017-07-24T00:11:02+01:00
-title: Documentation
-weight: 10
----
-
-
-<link rel="stylesheet" href="https://guangchuangyu.github.io/css/font-awesome.min.css">
-
-## <i class="fa fa-book"></i> Vignettes
-
-
-+ [Importing trees with data](http://bioconductor.org/packages/devel/bioc/vignettes/treeio/inst/doc/Importer.html)
-+ [Exporting trees with data](http://bioconductor.org/packages/devel/bioc/vignettes/treeio/inst/doc/Exporter.html)
-+ [ggtree for tree visualization and annotation](https://bioconductor.org/packages/devel/bioc/vignettes/ggtree/inst/doc/ggtree.html)
-
-
-## <i class="fa fa-wordpress"></i> Blog posts
-
-
-### <i class="fa fa-angle-double-right"></i> object conversion
-
-+ [convert graphic object to tree object using treeio](http://guangchuangyu.github.io/2017/01/convert-graphic-object-to-tree-object-using-treeio/)
-
