+++ site_src/content/index.rmd
@@ -1,128 +1,0 @@
----
-date: 2017-07-06T12:08:13+01:00
-title: 'treeio: Base classes and functions for parsing and exporting phylogenetic tree'
-type: index
-weight: 1
----
-
-
-<link rel="stylesheet" href="https://guangchuangyu.github.io/css/font-awesome.min.css">
-<link rel="stylesheet" href="https://guangchuangyu.github.io/css/academicons.min.css">
-
-
-<img src="https://raw.githubusercontent.com/Bioconductor/BiocStickers/master/treeio/treeio.png" height="200" align="right" />
-
-
-
-```{r echo=FALSE, results="hide", message=FALSE}
-library("ypages")
-```
-
-`r badge_bioc_release("treeio", "blue")`
-`r badge_devel("guangchuangyu/treeio", "blue")`
-`r badge_bioc_download("treeio", "total", "blue")`
-`r badge_bioc_download("treeio", "month", "blue")`
-
-
-
-The `treeio` package defines classes and functions for parsing and
-exporting phylogenetic trees.
-*treeio* is released within the [Bioconductor](https://bioconductor.org/packages/treeio/) project and the source code is hosted on <a href="https://github.com/GuangchuangYu/treeio"><i class="fa fa-github fa-lg"></i> GitHub</a>.
-
-
-## <i class="fa fa-user"></i> Authors
-
-Guangchuang Yu and Tommy Tsan-Yuk Lam, School of Public Health, The University of Hong Kong.
-
-<a href="https://twitter.com/guangchuangyu"><i class="fa fa-twitter fa-3x"></i></a>
-<a href="https://guangchuangyu.github.io/blog_images/biobabble.jpg"><i class="fa fa-wechat fa-3x"></i></a>
-<a href="https://www.ncbi.nlm.nih.gov/pubmed/?term=Guangchuang+Yu[Author+-+Full]"><i class="ai ai-pubmed ai-3x"></i></a>
-<a href="https://scholar.google.com.hk/citations?user=DO5oG40AAAAJ&hl=en"><i class="ai ai-google-scholar ai-3x"></i></a>
-<a href="https://orcid.org/0000-0002-6485-8781"><i class="ai ai-orcid ai-3x"></i></a>
-<a href="https://impactstory.org/u/0000-0002-6485-8781"><i class="ai ai-impactstory ai-3x"></i></a>
-
-
-## <i class="fa fa-book"></i> Citation
-
-`treeio` was splitted from `ggtree` package, please cite the following article when using `treeio`:
-
-`r badge_doi("10.1111/2041-210X.12628", "blue")`
-`r badge_altmetric("10533079", "blue")`
-`r badge_citation("HtEfBTGE9r8C", "7268358477862164627", "blue")`
-
-
-__G Yu__, DK Smith, H Zhu, Y Guan, TTY Lam<sup>\*</sup>. ggtree: an R package for visualization and annotation of phylogenetic trees with their covariates and other associated data. __*Methods in Ecology and Evolution*__. 2017, 8(1):28-36.
-
-<!--
-## <i class="fa fa-pencil"></i> Featured Articles
-
-![](https://guangchuangyu.github.io/featured_img/ggtree/2015_peiyu_1-s2.0-S1567134815300721-gr1.jpg)
-
-<i class="fa fa-hand-o-right"></i> Find out more on <i class="fa fa-pencil"></i> [Featured Articles](https://guangchuangyu.github.io/ggtree/featuredArticles/).
-
--->
-
-## <i class="fa fa-download"></i> Installation
-
-Install `treeio` is easy, follow the guide on the [Bioconductor page](https://bioconductor.org/packages/treeio/):
-
-```r
-## try http:// if https:// URLs are not supported
[…]
