+++ README.Rmd
@@ -91,16 +91,7 @@ devtools::install_github("GuangchuangYu/treeio")
 `r badge_bioc_download("treeio", "month", "blue")`
 
 
-```{r echo=F, comment=NA}
-x <- dlstats::bioc_stats("treeio")
-x <- x[-nrow(x), ] # last month data is not complete
-p <- ggplot(x, aes(end, Nb_of_distinct_IPs)) + geom_point() + geom_line() +
-    theme_minimal() +xlab(NULL) + ylab(NULL) +
-    labs(title="treeio", subtitle="Monthly download stats (distinct IPs)", caption="data from Bioconductor")
-ggsave(p, file="docs/images/dlstats.png", width=8, height=4)
-```
-
-<img src="docs/images/dlstats.png" width="890"/>
+<img src="https://guangchuangyu.github.io/software/treeio/index_files/figure-html/unnamed-chunk-2-1.png" width="890"/>
 
 
 ## Contributing
