### 2018-03-12 #179 @GuangchuangYu
@sckott thanks for accepting the pkg. All the requirements are incorporated.

----

@stefaniebutland I think for the first time of introducing a package, a long-form post is more appropriate.
