### 2018-03-15 #179 @stefaniebutland
Great to hear you're interested in contributing a post @GuangchuangYu! Here are some editorial and technical guidelines https://github.com/ropensci/roweb2#contributing-a-blog-post. You can have a look at other posts about onboarded software to see what style and length you like: https://ropensci.org/tags/review/

Would you be able to submit a draft via pull request 2018-04-03?  This gives us an opportunity to review it, give you feedback and confirm a publication date.
