+++ README.Rmd
@@ -52,7 +52,7 @@ associated data to a single tree file and can be served as a platform for
 merging tree with associated data and converting file formats.
 
 See the `treeio` project website, <https://guangchuangyu.github.io/software/treeio>, and
-package vignette for more details.
+package vignettes for more details.
 
 ### Vignettes
 
