### 2019-10-22 #179 @stefaniebutland
Congratulations @GuangchuangYu! I added the citation as a comment on your blog post https://ropensci.org/blog/2018/05/17/treeio/

(I used to dream of publishing in MBE back when I was doing research)
